### How the solution works

Generates a stream of numbers from `1` to `1_000_000`, filtering out only the primes.
With that subset, generate the number rotations and filter for circular primes only. Then count the remaining objects.
Circular number generation is done only after prime filtering for performance reasons.

### Performance

#### With prime filter

[Try it online!](https://tio.run/##dVPBbsIwDL3zFUaTUCJKaJl2QRo77cBpk3ZEqAqtC5HSlCUp64F/Z0mhUFo4NG387Gf72UUpKqFPp5fhtDR6uhFqiuoAWJsHKWZ5kZYS4VsXG4l5@PoGaTEAcMgehIn3WuRIIho48xwyLg320NkFtbrsgxWFvx0qqGABszM3wKcqc8al/CBkxpguSpWSec7tjplfbV0MdZSZIorCZAEac1IF4C7DdwgBVUodi3sNmmy6sNxiLIWxZLVDnsIRLBdyHUBTur/CeAw1vH4U6A@f5ezfRvpenlVNItqtwPW1Qe2rrQ2iUIY2TVf1eVzAUlncomap2AprCG3svZR1eDfKaXWLa4uwRRtfw0gcQBgAT5JLP@7rsaNXtu3YRydRAKteg24cRx@2pk94n/HVw0/czC3pKFH5sa/WN1XdFiVCJ6Xk@rZOXTnvE1z1uu3Y6LqMo4jeqwamkAdsKCPGojgMQ/80ND9WI89ZJqR1ffeonvh0qm47t9pvaqnrWX6xfWlN609kdXGD0@kf)

```
Real time: 11.570 s
User time: 11.377 s
Sys. time: 0.147 s
CPU share: 99.61 %
Exit code: 0
```

#### Without prime filter

[Try it online!](https://tio.run/##dVPBbsIwDL33K4yQUCJKaJl2QRo77cBpk3ZEqAqtgUhpypKU9cC/s6SjUFo4NG387Gf72UUpKqHP5@FgWho93Qg1RXUErM1Bhtu8yEqJ8KWLjcQ8enmFrAgAHHIAYZKDFjmSmIbOPIctlwZ76OyCWl32wYrC7x4VVLCA2T83wIcqc8alfCdkxpguSpWRec7tnpkfbV0MdZRbRRSFyQI05qQKwV0GbxABqow6FvcKmmy6sNxiIoWxZLVHnsEJLBdyHUJTur/CeAw1vH4U6A@f5d@/jfS9PKuaxLRbgetrg9pXWxtEoQxtmq7q87SApbK4Q80ysRPWENrYeynr8G6U0@oW1xZhhza5hpEkhCgEnqaXftzXY0evbNuxj07iEFa9Bt04Tj5sTZ/wPuOrh5@6mVvSUaLyY1@tb6q6LUqFTkvJ9W2dunLeJ7jqddux0XUZRzG9Vw1MIY/YUMaMxUkURf4Jhheeb6uR52wrpHWN97ie@HTKbju3@m@KqQtafrJDaU3rV2R1dcH5/Ac)

```
Real time: 18.005 s
User time: 17.591 s
Sys. time: 0.299 s
CPU share: 99.36 %
Exit code: 0
```